+++ libflv/include/mpeg4-avc.h
@@ -37,7 +37,7 @@ struct mpeg4_avc_t
 	uint8_t bit_depth_chroma_minus8;
 
     uint8_t data[4 * 1024];
-	int off;
+	size_t off;
 };
 
 int mpeg4_avc_decoder_configuration_record_load(const uint8_t* data, size_t bytes, struct mpeg4_avc_t* avc);
@@ -50,21 +50,21 @@ int mpeg4_avc_codecs(const struct mpeg4_avc_t* avc, char* codecs, size_t bytes);
 
 /// @param[out] vcl 0-non VCL, 1-IDR, 2-P/B
 /// @return <=0-error, >0-output bytes
-int h264_annexbtomp4(struct mpeg4_avc_t* avc, const void* data, int bytes, void* out, int size, int* vcl, int* update);
+int h264_annexbtomp4(struct mpeg4_avc_t* avc, const void* data, size_t bytes, void* out, size_t size, int* vcl, int* update);
 
 /// @return <=0-error, >0-output bytes
-int h264_mp4toannexb(const struct mpeg4_avc_t* avc, const void* data, int bytes, void* out, int size);
+int h264_mp4toannexb(const struct mpeg4_avc_t* avc, const void* data, size_t bytes, void* out, size_t size);
 
 /// h264_is_new_access_unit H.264 new access unit(frame)
 /// @return 1-new access, 0-not a new access
 int h264_is_new_access_unit(const uint8_t* nalu, size_t bytes);
 
 /// H.264 nal unit split
-int mpeg4_h264_annexb_nalu(const void* h264, int bytes, void (*handler)(void* param, const uint8_t* nalu, int bytes), void* param);
+int mpeg4_h264_annexb_nalu(const void* h264, size_t bytes, void (*handler)(void* param, const uint8_t* nalu, size_t bytes), void* param);
 
 /// Detect H.264 bitstrem type: H.264 Annexb or MP4-AVCC
 /// @return 0-annexb, >0-avcc length, <0-error
-int mpeg4_h264_bitstream_format(const uint8_t* h264, int bytes);
+int mpeg4_h264_bitstream_format(const uint8_t* h264, size_t bytes);
 
 #if defined(__cplusplus)
 }
